New enhancements
The code is now completely reviewed by Chris, Joshua, and Amir.
There seems to be a problem with data that causes ill-defined likelihood functions.
Chris is going to test whether a synthetiuc bivariate Gaussian Mixture in which the compoenents
are significantly far from each other resolves the problem or not.